fix(ui): break claude flags circular module dependency
