+++ src/ui/src/stores/slices/workspaceClaudeFlagsSlice.ts
@@ -4,13 +4,15 @@ import {
   type FlagValue,
   getResolvedRepoFlags,
 } from "../../services/claudeFlags";
+import {
+  hasDangerousFlag,
+  resolveEnabledExtraFlags,
+  type ResolvedFlag,
+} from "../../services/claudeFlagsLogic";
 import type { AppState } from "../useAppStore";
 
-export interface ResolvedFlag {
-  name: string;
-  value?: string;
-  isDangerous: boolean;
-}
+export type { ResolvedFlag };
+export { hasDangerousFlag, resolveEnabledExtraFlags };
 
 export type WorkspaceFlagsStatus = "loading" | "ready" | "error";
 
@@ -20,36 +22,6 @@ export interface WorkspaceFlagsState {
   repoState: Record<string, FlagValue>;
   resolved: ResolvedFlag[];
   status: WorkspaceFlagsStatus;
-}
-
-const DANGEROUS_FLAG = "--dangerously-skip-permissions";
-
-/// Mirror of `claudette::claude_flags_store::resolve_for_repo`. Repo state is
-/// the entries that have the `:override` sentinel set — the backend already
-/// filters by that, so any key present here wins over the matching global
-/// entry. Disabled flags are excluded; flags absent from `defs` are skipped;
-/// boolean flags emit `undefined` for value even if a stale value persists.
-export function resolveEnabledExtraFlags(
-  defs: ClaudeFlagDef[],
-  globalState: Record<string, FlagValue>,
-  repoState: Record<string, FlagValue>,
-): ResolvedFlag[] {
-  const out: ResolvedFlag[] = [];
-  for (const def of defs) {
-    const chosen = repoState[def.name] ?? globalState[def.name];
-    if (!chosen) continue;
-    if (!chosen.enabled) continue;
-    out.push({
-      name: def.name,
-      value: def.takes_value ? (chosen.value ?? "") : undefined,
-      isDangerous: def.is_dangerous,
-    });
-  }
-  return out;
-}
-
-export function hasDangerousFlag(resolved: ResolvedFlag[]): boolean {
-  return resolved.some((f) => f.name === DANGEROUS_FLAG);
 }
 
 export interface WorkspaceClaudeFlagsSlice {
